feat: initial logic for downloading oui files
There's likely a better way to do this...which means I'm off to a rough
start.  Currently, the server attempts to connect with `https` despite
the URL values being clearly declared as `http`.  Getting return codes
of `HTTP 418 - Unknown code`.  Will need to dig in to this one deeper.

Relates to #1